Add unit tests for transport counting functionality and moved class tests
